// models/pokeAPI.ts
export interface Species {
  name: string
  url: string
}

export interface Cries {
  latest: string
  legacy: string
}

export interface GameIndex {
  game_index: number
  version: Species
}

export interface NewMove {
  tag: string
  move: Species
  version_group_details: VersionGroupDetail[]
}

export interface VersionGroupDetail {
  level_learned_at: number
  move_learn_method: Species
  version_group: Species
}